requirements.txt
fastapi
pandas
uvicorn
numpy
scikit-learn
requests
torch
tqdm
python-dotenv
telethon
openai
tiktoken
xgboost
lightgbm
shap
optuna 